feat: 플레이어 공격 데미지 기능 추가
