+++ src/main/java/bossmonster/domain/attack/MagicDamageGenerator.java
@@ -1,9 +1,11 @@
-package bossmonster.domain;
+package bossmonster.domain.attack;
 
 public class MagicDamageGenerator implements DamageGenerator {
 
+    private static final int DAMAGE = 20;
+
     @Override
     public int generate() {
-        return 0;
+        return DAMAGE;
     }
 }
