+++ src/main/java/bossmonster/domain/attack/PhysicalDamageGenerator.java
@@ -1,9 +1,11 @@
-package bossmonster.domain;
+package bossmonster.domain.attack;
 
 public class PhysicalDamageGenerator implements DamageGenerator {
 
+    private static final int DAMAGE = 10;
+
     @Override
     public int generate() {
-        return 0;
+        return DAMAGE;
     }
 }
